separate the number spaces for contact, room and list handles...
everywhere where a handle has type flexibility now has a numeric handle type.
move the GetMembers and GetSelfHandle from Channel to Channel.Interface.Group,
and remove the Named interface in favour of a GetChannelHandle on channels
themselves. this is far more consistent between requesting, new channels
arriving, and inspecting existing channels, and removes the delusion that
it's a good idea for a UI that doesn't understand Group channels to handle
a channel that has a potentially mutable member list on protocols where
there is no concept of a one-to-one channel.

20051129185420-0e791-18c895ba49838f761ae484a34915f3ef041fc095.gz