Fixing TargetID handle search
It seems that there was a semantic error regarding TargetID. This patch
changes _alter_properties() to search for TargetID in each handle, that
would be the value returned by get_name(), which is different from
handle's id. The mistake was considering TargetID == handle.get_id()
in this function.
Another change was in EnsureChannel(). It was testing if the channel
existed before fixing the request.
There was a bug in _check_basic_properties(): line 436 was enforcing the
existance of a key that is allowed to be missing.

Signed-off-by: Thiago Borges Abdnur <bolaum@gmail.com>
Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>