Make sure to create the _generated directory before creating __init__.py.
When building in parallel, the xsltproc processes and the single echo
process will be most likely simultaneously (for an high enough amount of
jobs).

The echo is straightforward, while the XSLT processing takes time, this
means that it'd be the first one to complete and thus create its output;
unfortunately while xsltproc does create the structure of directories if it
doesn't find it already, shell output redirection won't, so you have to
create it beforehand for the echo to complete properly.

An easy way to test this before was to simply request
“make _generated/__init__.py” on a pristine source directory.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>