Fix CreateChannel to always return a new channel.
Fix CreateChannel to always return a new channel instead of having the
same behaviour as EnsureChannel and so allow to have multiple channel of
the same type created for a given TargetHandle. Provide a existing_channel
function in ChanelManager that ConnectionManager can subclass to define
more appropriatly if there is an existing channel that match the requested
properties.