Fix registering new channels in channel manager
Fix registering new channels in channel manager, thanks Edward Page for
noticing.
Checking for the handle to be in the dictionary  contradict using setdefault.
The handle never ends up in self._channels[type] and so the conditional always
fails so the channel never ends up being added to the list.